Fixes:
As it turned out - there were retain cycles all over BLEOperationsManager due to incorrect usage of '__block' and total absence of 'operations' array initialization. This was a bad thing - and it was fixed.

I removed code that was saving and restoring centralManager delegate. It was a bad idea, as when I chained operations and some operation in chain tried to restore a delegate back, the previous operation (to which was that weak ref pointing) had been already deallocated, thus - a crash.

I put some weak refs into blocks to prevent retain cycles - though those blocks may not cause cycles right now.